test: fix leftover route from test_add_remove_routes
Commit f8c57a88fe05 ('route: Add support for the `route-type`') added a
blackhole route to the example `eth1_add_route.yml`. This is not very
correct because the route is not related to eth1, as the rest of the
example.

Moreover, this example is used in some integration tests. The example to
do the cleanup was not modified to clean the blackhole too. This causes
that the blackhole route is not cleaned after the test.

Move the blackhole route to its own example and test, doing proper
cleanup. This also fixes some CI failures caused by this leftover route.
The end cause of those CI failures is a bug in NetworkManager
(https://issues.redhat.com/browse/RHEL-71686), but we are still suposed
to do proper cleanup after each test, anyway.

Signed-off-by: Íñigo Huguet <ihuguet@redhat.com>